clippy +fmt

+++ src/merge_variants.rs
@@ -59,7 +59,6 @@ where
         }
 
         if line.starts_with("SQ") {
-
             seq_len = line[13..]
                 .split_whitespace()
                 .next()
@@ -80,8 +79,8 @@ where
             if let Some(last_aa_valid) = last_aa {
                 for candidate_slice in insert_candidates.drain(..) {
                     for mut candidate in candidate_slice.iter().cloned() {
-                        candidate.normalize(seq_len as u32, last_aa_valid);
-                        if !variants_in_entry.contains(&candidate) || confirmed_only{
+                        candidate.normalize(seq_len, last_aa_valid);
+                        if !variants_in_entry.contains(&candidate) || confirmed_only {
                             write!(writer, "{}", candidate.to_uniprot(seq_len)?)?;
                         }
                     }
@@ -127,9 +126,7 @@ where
                         //CASE: aa_change line complete -> insert complete entry
                         if segment.starts_with("/") {
                             let note = aa_change.get(28..).unwrap();
-                            if let Some(caps) =
-                                regex_aa_change.captures(note)
-                            {
+                            if let Some(caps) = regex_aa_change.captures(note) {
                                 let aa_ref = &caps[1];
                                 let aa_new = &caps[2];
 
@@ -139,7 +136,7 @@ where
                                     aa_ref: aa_ref.to_string(),
                                     aa_new: aa_new.to_string(),
                                 });
-                            } else if note.starts_with("Missing"){
+                            } else if note.starts_with("Missing") {
                                 variants_in_entry.push(Variant {
                                     pos_start: pos.0,
                                     pos_end: pos.1,
@@ -160,7 +157,6 @@ where
                         aa_change_line = Some(line);
                     }
                 }
-
                 //CASE: Check if FT line containing VARIANT or VAR_SEQ. Sets aa_change_pos
                 else {
                     let mut parts = line
